feat: refactor cart and checkout
