+++ client-app/pages/checkout/billing.vue
@@ -19,8 +19,8 @@
                 class="px-3 self-start uppercase font-bold"
                 @click="
                   addresses.length
-                    ? openAddressSelectionDialog()
-                    : openAddOrUpdateAddressDialog(payment?.billingAddress)
+                    ? openSelectAddressModal(AddressType.Billing)
+                    : openAddOrUpdateAddressModal(AddressType.Billing, payment?.billingAddress)
                 "
               >
                 {{ $t("pages.checkout.shipping_details_section.shipping_address_block.change_button") }}
@@ -49,7 +49,11 @@
                   size="sm"
                   is-outline
                   class="px-3 self-start uppercase font-bold"
-                  @click="addresses.length ? openAddressSelectionDialog() : openAddOrUpdateAddressDialog()"
+                  @click="
+                    addresses.length
+                      ? openSelectAddressModal(AddressType.Billing)
+                      : openAddOrUpdateAddressModal(AddressType.Billing, payment?.billingAddress)
+                  "
                 >
                   {{ $t("pages.checkout.payment_details_section.billing_address_block.add_address_button") }}
                 </VcButton>
@@ -76,7 +80,7 @@
                 size="sm"
                 is-outline
                 class="px-3 self-start uppercase font-bold"
-                @click="showPaymentMethodDialog"
+                @click="openSelectPaymentMethodModal"
               >
                 {{
                   payment?.paymentGatewayCode
@@ -93,104 +97,11 @@
 </template>
 
 <script setup lang="ts">
-import { computed, onMounted } from "vue";
-import { omit } from "lodash";
-import {
-  AddOrUpdateAddressDialog,
-  CheckoutLabeledBlock,
-  PaymentMethodDialog,
-  SelectAddressDialog,
-} from "@/shared/checkout";
-import { usePopup } from "@/shared/popup";
-import { useCart } from "@/shared/cart";
-import { useUser, useUserAddresses, useUserCheckoutDefaults } from "@/shared/account";
-import { CartAddressType, InputAddressType, MemberAddressType, PaymentMethodType, PaymentType } from "@/xapi";
-import { AddressType, convertToType } from "@/core";
+import { AddressType } from "@/core";
+import { useUser, useUserAddresses } from "@/shared/account";
+import { CheckoutLabeledBlock, useCheckout } from "@/shared/checkout";
 
 const { user, isAuthenticated } = useUser();
-const { cart, fetchCart, updatePayment } = useCart();
 const { addresses } = useUserAddresses({ user });
-const { openPopup, closePopup } = usePopup();
-const { getUserCheckoutDefaults } = useUserCheckoutDefaults();
-
-const { paymentMethodCode } = getUserCheckoutDefaults() ?? {};
-
-const availablePaymentMethods = computed<PaymentMethodType[]>(() => cart.value.availablePaymentMethods ?? []);
-const defaultPaymentMethod = availablePaymentMethods.value.find(
-  (item: PaymentMethodType) => item.code === paymentMethodCode
-);
-const payment = computed<PaymentType | undefined>(() => cart.value.payments?.[0]);
-
-function openAddressSelectionDialog(): void {
-  openPopup({
-    component: SelectAddressDialog,
-    props: {
-      addresses: addresses.value,
-      currentAddress: convertToType<MemberAddressType>(payment.value?.billingAddress),
-
-      async onResult(selectedAddress: MemberAddressType): Promise<void> {
-        selectedAddress.addressType = AddressType.Billing;
-        await updatePayment({
-          id: payment.value?.id,
-          billingAddress: convertToType<InputAddressType>(omit(selectedAddress, ["isDefault", "description"])),
-        });
-        closePopup();
-      },
-
-      onAddNewAddress() {
-        setTimeout(() => {
-          openAddOrUpdateAddressDialog();
-        }, 500);
-      },
-    },
-  });
-}
-
-function openAddOrUpdateAddressDialog(currentAddress?: CartAddressType): void {
-  openPopup({
-    component: AddOrUpdateAddressDialog,
-    props: {
-      address: currentAddress,
-
-      async onResult(updatedAddress: MemberAddressType): Promise<void> {
-        updatedAddress.addressType = AddressType.Billing;
-        await updatePayment({
-          id: payment.value?.id,
-          billingAddress: convertToType<InputAddressType>(omit(updatedAddress, ["isDefault", "description"])),
-        });
-        closePopup();
-      },
-    },
-  });
-}
-
-function showPaymentMethodDialog(): void {
-  openPopup({
-    component: PaymentMethodDialog,
-    props: {
-      currentMethodCode: payment.value?.paymentGatewayCode,
-      availableMethods: availablePaymentMethods.value,
-      async onResult(method: PaymentMethodType) {
-        await updatePayment({
-          id: payment.value?.id,
-          paymentGatewayCode: method.code,
-        });
-      },
-    },
-  });
-}
-
-onMounted(async () => {
-  if (!payment.value?.paymentGatewayCode && paymentMethodCode && defaultPaymentMethod) {
-    await updatePayment(
-      {
-        id: payment.value?.id,
-        paymentGatewayCode: defaultPaymentMethod.code,
-      },
-      false
-    );
-
-    await fetchCart();
-  }
-});
+const { payment, openSelectAddressModal, openAddOrUpdateAddressModal, openSelectPaymentMethodModal } = useCheckout();
 </script>
