+++ client-app/shared/checkout/components/index.ts
@@ -1,7 +1,7 @@
 export { default as CheckoutLabeledBlock } from "./checkout-labeled-block.vue";
 export { default as OrderSummary } from "./order-summary.vue";
-export { default as ShippingMethodDialog } from "./shipping-method-dialog.vue";
-export { default as PaymentMethodDialog } from "./payment-method-dialog.vue";
-export { default as SelectAddressDialog } from "./select-address-dialog.vue";
-export { default as AddOrUpdateAddressDialog } from "./add-or-update-address-dialog.vue";
+export { default as SelectShippingMethodModal } from "./select-shipping-method-modal.vue";
+export { default as SelectPaymentMethodModal } from "./select-payment-method-modal.vue";
+export { default as SelectAddressModal } from "./select-address-modal.vue";
+export { default as AddOrUpdateAddressModal } from "./add-or-update-address-modal.vue";
 export { default as AcceptedGifts } from "./accepted-gifts.vue";
