+++ client-app/shared/checkout/composables/index.ts
@@ -1,0 +1,1 @@
+export { default as useCheckout } from "./useCheckout";
