+++ client-app/shared/checkout/index.ts
@@ -1,2 +1,3 @@
 export * from "./components";
+export * from "./composables";
 export * from "./utilities";
